+++ src/core/config/config_struct.rs
@@ -6,7 +6,7 @@ use crate::{
         Alias, AnyControlMsg, ControlFunction, ControlMsg, DataControlMsg,
         config::default_http_client_path, player::AddPolicy, query::Query,
     },
-    env::install,
+    env::{RunType, install},
     ext::Wrap,
     gen_struct,
 };
@@ -125,6 +125,10 @@ gen_struct! {
             default_http_client_path()
         },
 
+        web_client_command: Option<String> { pub, pub } => pub(super) () {
+            None
+        },
+
         ; // fields passed by value:
 
         #[doc = "When enabled uamp will sort only by the primary attribute."]
@@ -209,6 +213,11 @@ gen_struct! {
         #[doc = "changes."]
         auto_restart: bool { pub, pub } => pub(super) () true,
 
+        #[doc = "The run type to use by default."]
+        default_run_type: RunType { pub, pub } => pub(super) () {
+            RunType::WebClient
+        },
+
         ; // fields that aren't serialized
 
         #[serde(skip_serializing, default = "default_config_path_json")]
@@ -269,6 +278,8 @@ impl Config {
             auto_restart: default_auto_restart(),
             force_server: None,
             skin: default_skin(),
+            default_run_type: default_default_run_type(),
+            web_client_command: default_web_client_command(),
             change: Cell::new(true),
         }
     }