+++ src/env/mod.rs
@@ -2,10 +2,11 @@ mod app_ctrl;
 mod command;
 pub mod install;
 pub mod rt;
+mod run_type;
 mod streams;
 
 //===========================================================================//
 //                                   Public                                  //
 //===========================================================================//
 
-pub use self::{app_ctrl::*, command::*};
+pub use self::{app_ctrl::*, command::*, run_type::*};