+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - Allow specifying cnt for `pop`.
 - Add option to check if self update is enabled with `update --enabled`.
 - Add new `internal` action `install`.
+- Add option to open the web client with `uamp run -w`.
+- Add option to configure the default run type when running uamp without
+  arguments with `default_run_type`.
+- Add option to specify the web client command with `web_client_command`.
+- Add new internal action `open`.
+- Uamp now also installs web client, icons and desktop entry.
 
 ### Fixes
 - Fix order of `a` and `b` when requesting with `uam i nfo=a..b`.